Task2/requirements.txt
torch
numpy~=1.8
scikit-learn
torchvision
pillow
datasets
transformers
pandas
contractions
nltk
seqeval